requirements.txt
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
airflow-code-editor==7.0.1
alembic==1.8.1
anyio==3.6.1
apache-airflow==2.3.4
apache-airflow-providers-common-sql==1.1.0
apache-airflow-providers-ftp==3.1.0
apache-airflow-providers-http==4.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-sqlite==3.2.0
apispec==3.3.2
appdirs==1.4.4
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bleach==6.0.0
blinker==1.5
cachelib==0.9.0
cattrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
clickclick==20.10.2
cloudpickle==2.2.1
colorama==0.4.5
colorlog==4.8.0
commonmark==0.9.1
connexion==2.14.1
cron-descriptor==1.2.31
croniter==1.3.7
cryptography==36.0.2
cycler==0.11.0
dask==2023.2.1
datadog==0.44.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
dnspython==2.2.1
docker==6.0.1
docutils==0.19
email-validator==1.2.1
entrypoints==0.4
exceptiongroup==1.0.0rc9
executing==1.0.0
fabulous==0.4.0
Faker==4.14.0
fastjsonschema==2.16.3
findspark==2.0.1
Flask==2.2.2
Flask-AppBuilder==4.1.3
Flask-Babel==2.0.0
Flask-Caching==2.0.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
fonttools==4.37.1
freezegun==1.2.2
frozenlist==1.3.3
fs==2.4.16
fsspec==2023.1.0
graphviz==0.20.1
greenlet==1.1.3
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.1
idna==3.3
importlib-metadata==4.12.0
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.15.0
ipython==8.4.0
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.15.0
jupyter-client==7.3.4
jupyter-server==1.23.5
jupyter-server-proxy==3.2.1
jupyter_core==5.2.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
linkify-it-py==2.0.0
locket==1.0.0
lockfile==0.12.2
lxml==4.9.1
mage-ai==0.8.7
Mako==1.2.2
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.17.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
matplotlib-venn==0.11.7
mdit-py-plugins==0.3.0
mdurl==0.1.2
minio==7.1.11
mistune==2.0.5
multidict==6.0.4
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.5
numpy==1.23.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.0.0
pluggy==1.0.0
prison==0.2.1
prometheus-client==0.16.0
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
pyarrow==10.0.1
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.18.1
pyspark==3.3.2
pytest==7.2.2
python-daemon==2.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-nvd3==0.15.0
python-slugify==6.1.2
pytz==2022.1
pytzdata==2020.1
PyYAML==6.0
pyzmq==23.2.1
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==12.5.1
scikit-learn==1.2.1
scipy==1.9.1
seaborn==0.11.2
Send2Trash==1.8.0
setproctitle==1.3.2
shellingham==1.5.0.post1
simpervisor==0.4
simplejson==3.18.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.3
sqlparse==0.4.2
stack-data==0.5.0
swagger-ui-bundle==0.0.9
tabulate==0.8.10
tenacity==8.0.1
termcolor==1.1.0
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
typer==0.7.0
typing_extensions==4.3.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==2.3.3
yarl==1.8.2
zipp==3.8.1
